rag/pyproject.toml
[project]
name = "legalbot-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan", email = "iskiba60@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0,<0.122.0",
    "uvicorn>=0.38.0,<0.39.0",
    "numpy>=2.0.0,<3.0.0",
    "pydantic>=2.12.4,<3.0.0",
    "pydantic-settings>=2.9.0,<3.0.0",
    "openai>=1.0.0,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "pgvector>=0.3.6,<1.0.0",
    "pymupdf>=1.26.0,<2.0.0",
    "sse-starlette>=2.3.0,<3.0.0",
    "tenacity>=9.1.0,<10.0.0",
    "httpx>=0.28.0,<1.0.0",
    "python-docx>=1.1.0,<1.2.0",
    "python-multipart>=0.0.9,<0.1.0",
    "striprtf>=0.0.28"
]

[tool.poetry]
name = "legalbot-rag"
version = "0.1.0"
description = ""
authors = ["Ivan <iskiba60@gmail.com>"]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.uv]
package = false
